// src/typings/tapi/saleToPOIRequest.ts
export class SaleToPOIRequest {
    "MessageHeader": MessageHeader;
// ...
}